# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Focus",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://dev.api.focus.grupomobius.com"
        },
        {
            "url": "https://api.focus.grupomobius.com"
        }
    ],
    "paths": {
...
        "/api/purchase-operations/getall": {
            "get": {
                "tags": [
                    "purchase-operations"
                ],
                "summary": "Get all purchase operatons",
                "operationId": "257c5ac9223081fdae1d3b2ae13df39d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseOperation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/purchase-operations/{id}": {
            "get": {
                "tags": [
                    "purchase-operations"
                ],
                "summary": "Get purchase operation by ID",
                "operationId": "c502449c88d9a3e692bcb26a45c0cfe8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseOperation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Purchase Operation not found."
                    }
                }
            }
        },
        "/purchase-operations/update/{id}": {
            "put": {
                "tags": [
                    "purchase-operations"
                ],
                "summary": "Updated purchase operation",
                "operationId": "updatePurchaseOperation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated purchase operation object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseOperation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseOperation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Purchase operation not found"
                    }
                }
            }
        },
...
        "/api/questions/getall": {
            "get": {
                "tags": [
                    "questions"
                ],
                "summary": "Get all type questions",
                "operationId": "4d9354eafb654909f0747d8e8705d498",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
...
        "/api/reception/{id}": {
            "get": {
                "tags": [
                    "receptions"
                ],
                "summary": "Get reception by ID",
                "operationId": "42acd3b0bad65239817b1d78b3629310",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reception"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reception not found."
                    }
                }
            }
        },
        "reception/{id}": {
            "put": {
                "tags": [
                    "receptions"
                ],
                "summary": "Updated reception",
                "operationId": "updateReception",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated reception object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reception"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reception"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reception not found"
                    }
                }
            }
        },
...
        "/api/state-pending-tasks/getall": {
            "get": {
                "tags": [
                    "state-pending-tasks"
                ],
                "summary": "Get all state pending tasks",
                "operationId": "b394328066159fde44ff9dbe36efd135",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatePendingTask"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/state-pending-tasks/{id}": {
            "get": {
                "tags": [
                    "state-pending-tasks"
                ],
                "summary": "Get state pending task by ID",
                "operationId": "614d75d6b9f0e7f791f163cf740af338",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatePendingTask"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "State Pending Task not found."
                    }
                }
            }
        },
        "/state-pending-tasks/update/{id}": {
            "put": {
                "tags": [
                    "state-pending-tasks"
                ],
                "summary": "Updated purchase state pending task",
                "operationId": "updateStatePendingTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated purchase operation object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatePendingTask"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatePendingTask"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "State pending task not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PurchaseOperation": {
                "title": "Purchase Operation model",
                "description": "Purchase Operation model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "task_id": {
                        "title": "Task ID",
                        "description": "Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Price",
                        "type": "number",
                        "format": "double"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Question": {
                "title": "Question model",
                "description": "Question model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "company_id": {
                        "title": "Company ID",
                        "description": "Company ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "question": {
                        "title": "Question",
                        "description": "Question",
                        "type": "string"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Reception": {
                "title": "Reception model",
                "description": "Reception model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "campa_id": {
                        "title": "Campa ID",
                        "description": "Campa ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicle_id": {
                        "title": "Vehicle ID",
                        "description": "Vehicle ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "type_model_order_id": {
                        "title": "Type of Model Order ID",
                        "description": "Type of Model Order ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "group_task_id": {
                        "title": "Group Task ID",
                        "description": "Group Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "finished": {
                        "title": "Finished",
                        "description": "Finished",
                        "type": "boolean"
                    },
                    "has_accessories": {
                        "title": "Has accesories",
                        "description": "Has accesories",
                        "type": "boolean"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "When was deleted",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "StatePendingTask": {
                "title": "State Pending Task model",
                "description": "State Pending Task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}